\encoding{UTF-8}
\name{tkRplotR-package}
\alias{tkRplotR-package}
\alias{tkRplotR}
\docType{package}
\title{Display Resizable Plots}
\description{
  This package contains functions for ploting in a Tk canvas.
}
\details{
\tabular{ll}{
Package: \tab tkRplotR\cr
Type: \tab Package\cr
License: \tab \acronym{GPL} (>= 2)\cr
}
\emph{Main Functions}
  \describe{
    \item{\code{\link{tkRplot}}}{ display a plot in a Tk toplevel window }
    \item{\code{\link{tkRreplot}}}{ refresh the plot created by tkRplot }
  }
}
\author{
  Filipe Campelo \email{fcampelo@ci.uc.pt}
  }
\keyword{ package }
