% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setVariable}
\alias{setVariable}
\alias{getVariable}
\title{Set and Get Variables}
\usage{
setVariable(name, value = NULL)
getVariable(name)
}
\arguments{
\item{name}{name of the variable}

\item{value}{the value of the variable}
}
\description{
Define and get variables
}
\examples{
setVariable("var1", 1)
exists("var1")
getVariable("var1")

getVariable("tkRplotR_pngType")

}
